+++ packages/winnow/test/helper.js
@@ -1,0 +1,10 @@
+const winnow = require('../src')
+
+module.exports = {
+  run (data, options, expected, t) {
+    t.plan(1)
+    const features = require(`./fixtures/${data}.json`).features
+    const filtered = winnow.query(features, options)
+    t.equal(filtered.length, expected)
+  }
+}